avx512: [zuc] ZUC intercept optimizations
- SAFE_DATA mask compute optimized
- Eliminated one bit-test operation
- "_zuc_init_not_done", "_zuc_unused_lane_bitmask",
  "_zuc_completed_job_bitmask" and "_zuc_fresh_init" changed from 16-bit
  wide to 32-bit wide
- index_to_mask table converted from 16-bit to 32-bits
- re-organized the module to make code and data sections more visible
- replaced broadcast-word with some alternatives

+++ lib/include/ipsec_ooo_mgr.h
@@ -354,10 +354,10 @@ typedef struct {
         IMB_JOB *job_in_lane[16];
         uint64_t num_lanes_inuse;
         DECLARE_ALIGNED(uint32_t state[MAX_ZUC_STATE_SZ], 64);
-        uint16_t init_not_done;
-        uint16_t unused_lane_bitmask;
-        uint16_t completed_job_bitmask;
-        uint16_t fresh_init;
+        uint32_t init_not_done;
+        uint32_t unused_lane_bitmask;
+        uint32_t completed_job_bitmask;
+        uint32_t fresh_init;
         uint64_t road_block;
 } MB_MGR_ZUC_OOO;
